refactored data to use hook
